I2C fixes and improvements

+++ TactilityC/Source/tt_hal_i2c.cpp
@@ -23,11 +23,11 @@ bool tt_hal_i2c_master_read_register(i2c_port_t port, uint8_t address, uint8_t r
     return tt::hal::i2c::masterReadRegister(port, address, reg, data, dataSize, timeout);
 }
 
-bool tt_hal_i2c_master_write(i2c_port_t port, uint16_t address, const uint8_t* data, uint16_t dataSize, TickType_t timeout) {
+bool tt_hal_i2c_master_write(i2c_port_t port, uint8_t address, const uint8_t* data, uint16_t dataSize, TickType_t timeout) {
     return tt::hal::i2c::masterWrite(port, address, data, dataSize, timeout);
 }
 
-bool tt_hal_i2c_master_write_register(i2c_port_t port, uint16_t address, uint8_t reg, const uint8_t* data, uint16_t dataSize, TickType_t timeout) {
+bool tt_hal_i2c_master_write_register(i2c_port_t port, uint8_t address, uint8_t reg, const uint8_t* data, uint16_t dataSize, TickType_t timeout) {
     return tt::hal::i2c::masterWriteRegister(port, address, reg, data, dataSize, timeout);
 }
